+Component Constructor,+=Generic Dictionary,+Modular Input System
The generation of Components is now done using the new Component Constructor. The Component Constructor is used to create and modify Components that act as children an class that implements it.

Generic Dictionaries can no longer have null values.

The Modular Input system allows the user to define Inputs using an Input Asset and tell the Player Controller Scriptable Object what Method to call for those buttons presses.