Major Refactor, +Dialogue System, +Bug Fixes, +More
The previous update of the CCDK had a breaking bug, rendering it impossible to use. This has been fixed. Also, the Toolkit is now released via a UnityPackage so that required depencies are installed when it's included in projects.

The Dialogue System is a power and extendable tool for displaying text (typically speech) to the player. It can be used in many ways, including Bolt and Timeline.